// src/app/services/poke.interfaces.ts
export interface PokemonInfo {
  name?: string;
  id?: number;
  sprites: {
    front_default: string;
  };
  types?: Type[];
  moves?: move[];
}

export interface move {
  name?: string;
  url?: string
}

export interface Type {
  slot?: number;
  type?: {
    name?: string;
    url?: string;
  };
}
export interface PokemonCaracts {
  descriptions: {
    description: string;
    language: {
      name: string;
      url: string;
    };
  }[];
}